+++ NEWS.md
@@ -1,6 +1,7 @@
 # qualtRics (development version)
 
 - Add `fetch_description()` to download complete survey description metadata from v3 API endpoint (more up-to-date than older `metadata()`) thanks to @jmobrien (#207)
+- Warn users about possible incorrect results from API when `breakout_sets` and `label` are both FALSE
 
 # qualtRics 3.1.4